Fetching data for each Page correctly.
